fix: User 및 Match entity FK 타입 수정, bigint 비교 로직 추가
- User 및 Match entity에서 FK 필드 타입을 `int4` → `bigint`로 수정
- TypeORM bigint 반환값을 string에서 Number로 변환하여 비교하도록 로직 추가

+++ packages/backend/src/match/entity/match.entity.ts
@@ -13,16 +13,16 @@ import { UserProblemBank } from '../../problem-bank/entity';
 
 @Entity('matches')
 export class Match {
-  @PrimaryGeneratedColumn('increment', { type: 'int4' })
+  @PrimaryGeneratedColumn('increment', { type: 'bigint' })
   id: number;
 
-  @Column({ type: 'int4', nullable: false, name: 'player1_id' })
+  @Column({ type: 'bigint', nullable: false, name: 'player1_id' })
   player1Id: number;
 
-  @Column({ type: 'int4', nullable: true, name: 'player2_id' })
+  @Column({ type: 'bigint', nullable: true, name: 'player2_id' })
   player2Id: number | null;
 
-  @Column({ type: 'int4', nullable: true, name: 'winner_id' })
+  @Column({ type: 'bigint', nullable: true, name: 'winner_id' })
   winnerId: number | null;
 
   @Column({
