+++ packages/backend/src/user/entity/user.entity.ts
@@ -13,7 +13,7 @@ import { UserProblemBank } from '../../problem-bank/entity';
 
 @Entity('users')
 export class User {
-  @PrimaryGeneratedColumn('increment', { type: 'int4' })
+  @PrimaryGeneratedColumn('increment', { type: 'bigint' })
   id: number;
 
   @Column({ type: 'varchar', nullable: true })
